\name{SEMundir-package}
\alias{SEMundir-package}
\alias{SEMundir}
\docType{package}
\title{
\packageTitle{SEMundir}
}
\description{
\packageDescription{SEMundir}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SEMundir}
\packageIndices{SEMundir}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SEMundir}

Maintainer: \packageMaintainer{SEMundir}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
